NPC 랜덤 무브 도중 길 탐색 시 바로 추적 무브로 할 수 있게 변경, NPC리스폰 시 뷰 리스트 업데이트 플레이어 버전으로 업데이트 되는 부분 수정

+++ Terproject/Terproject/GameObject/NPC_Object/NPC_Object.cpp
@@ -1,12 +1,11 @@
 #include "stdafx.h"
 #include "NPC_Object.h"
-#include "NPC_Func.h"
 #include "../../Logic/Logic.h"
 
 extern random_device g_rd;
 extern default_random_engine g_dre;
 extern uniform_int_distribution<int> g_npcRandDir; // inclusive
-extern uniform_int_distribution<int> g_npcRandPostion; // inclusive
+extern uniform_int_distribution<short> g_npcRandPostion; // inclusive
 
 
 NPC_Object::NPC_Object() : GameObject()
@@ -34,7 +33,6 @@ NPC_Object::~NPC_Object()
 void NPC_Object::SetRandPosition()
 {
 	SetPosition(g_npcRandPostion(g_dre), g_npcRandPostion(g_dre));
-	//GameMap¿¡ ¾ê³× ¾È³Ö¾îÁÜ.
 }
 
 void NPC_Object::RemoveViewListPlayer(int removePlayerId)
